tools/VINS-Configurator/requirements.txt
action-msgs==1.2.1
action-tutorials-interfaces==0.20.3
action-tutorials-py==0.20.3
actionlib-msgs==4.2.3
ament-cmake-test==1.3.8
ament-copyright==0.12.10
ament-cppcheck==0.12.10
ament-cpplint==0.12.10
ament-flake8==0.12.10
ament-index-python==1.4.0
ament-lint==0.12.10
ament-lint-cmake==0.12.10
ament-package==0.14.0
ament-pep257==0.12.10
ament-uncrustify==0.12.10
ament-xmllint==0.12.10
angles==1.15.0
blinker==1.8.2
builtin-interfaces==1.2.1
click==8.1.7
composition-interfaces==1.2.1
cv-bridge==3.2.1
demo-nodes-py==0.20.3
diagnostic-msgs==4.2.3
domain-coordinator==0.10.0
example-interfaces==0.9.3
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
Flask==3.0.3
geometry-msgs==4.2.3
image-geometry==3.2.1
interactive-markers==2.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
laser-geometry==2.4.0
launch==1.0.5
launch-ros==0.19.7
launch-testing==1.0.5
launch-testing-ros==0.19.7
launch-xml==1.0.5
launch-yaml==1.0.5
lifecycle-msgs==1.2.1
logging-demo==0.20.3
map-msgs==2.1.0
MarkupSafe==2.1.5
message-filters==4.3.3
nav-msgs==4.2.3
osrf-pycommon==2.0.2
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
python-qt-binding==1.1.2
PyYAML==6.0.2
qt-dotgraph==2.2.3
qt-gui==2.2.3
qt-gui-cpp==2.2.3
qt-gui-py-common==2.2.3
quality-of-service-demo-py==0.20.3
rcl-interfaces==1.2.1
rclpy==3.3.12
rcutils==5.1.5
resource-retriever==3.1.1
rmw-dds-common==1.6.0
ros2action==0.18.9
ros2bag==0.15.9
ros2cli==0.18.9
ros2component==0.18.9
ros2doctor==0.18.9
ros2interface==0.18.9
ros2launch==0.19.7
ros2lifecycle==0.18.9
ros2multicast==0.18.9
ros2node==0.18.9
ros2param==0.18.9
ros2pkg==0.18.9
ros2run==0.18.9
ros2service==0.18.9
ros2topic==0.18.9
rosbag2-interfaces==0.15.9
rosbag2-py==0.15.9
rosgraph-msgs==1.2.1
rosidl-adapter==3.1.5
rosidl-cli==3.1.5
rosidl-cmake==3.1.5
rosidl-generator-c==3.1.5
rosidl-generator-cpp==3.1.5
rosidl-generator-py==0.14.4
rosidl-parser==3.1.5
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.1
rosidl-typesupport-cpp==2.0.1
rosidl-typesupport-fastrtps-c==2.2.2
rosidl-typesupport-fastrtps-cpp==2.2.2
rosidl-typesupport-introspection-c==3.1.5
rosidl-typesupport-introspection-cpp==3.1.5
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.4
rqt-bag-plugins==1.1.4
rqt-console==2.0.3
rqt-graph==1.3.0
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-common==1.1.7
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
sensor-msgs==4.2.3
sensor-msgs-py==4.2.3
shape-msgs==4.2.3
sros2==0.10.4
statistics-msgs==1.2.1
std-msgs==4.2.3
std-srvs==4.2.3
stereo-msgs==4.2.3
teleop-twist-keyboard==2.4.0
tf2-geometry-msgs==0.25.6
tf2-kdl==0.25.6
tf2-msgs==0.25.6
tf2-py==0.25.6
tf2-ros-py==0.25.6
tf2-tools==0.25.6
topic-monitor==0.20.3
trajectory-msgs==4.2.3
turtlesim==1.4.2
unique-identifier-msgs==2.2.1
visualization-msgs==4.2.3
Werkzeug==3.0.3
